feat!: mark all framework plugins as optional peer
