+++ .changeset/curvy-symbols-type.md
@@ -1,0 +1,5 @@
+---
+"@1stg/eslint-config": major
+---
+
+feat!: mark all framework plugins as optional peer
