+++ package.json
@@ -28,9 +28,12 @@
     "@angular/compiler": "^19.2.6",
     "@angular/compiler-cli": "^19.2.6",
     "@angular/core": "^19.2.6",
+    "@babel/eslint-parser": "^7.27.0",
+    "@babel/eslint-plugin": "^7.27.0",
     "@changesets/changelog-github": "^0.5.1",
     "@changesets/cli": "^2.28.1",
     "@commitlint/cli": "^19.8.0",
+    "@eslint-react/eslint-plugin": "^1.47.0",
     "@types/babel__core": "^7.20.5",
     "@types/bun": "^1.2.9",
     "@types/node": "^22.14.1",
@@ -38,8 +41,15 @@
     "@types/unist": "^3.0.3",
     "@unts/patch-package": "^8.1.1",
     "@vitest/coverage-istanbul": "^3.1.1",
+    "@vitest/eslint-plugin": "^1.1.42",
+    "angular-eslint": "^19.3.0",
+    "angular-eslint-template-parser": "^0.1.3",
     "eslint": "^9.24.0",
     "eslint-config-prettier": "^10.1.2",
+    "eslint-plugin-jest": "^28.11.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-vue": "^10.0.0",
+    "jest": "^29.7.0",
     "lint-staged": "^15.5.1",
     "markuplint": "^4.11.8",
     "npm-run-all2": "^7.0.2",
