+++ packages/eslint-config/package.json
@@ -21,26 +21,57 @@
     "./recommended": "./recommended.js"
   },
   "peerDependencies": {
-    "eslint": ">=8.0.0"
+    "@babel/eslint-parser": "^7.0.0",
+    "@babel/eslint-plugin": "^7.0.0",
+    "@eslint-react/eslint-plugin": "^1.0.0",
+    "@vitest/eslint-plugin": "^1.0.0",
+    "angular-eslint": "^19.0.0",
+    "angular-eslint-template-parser": "^0.1.0",
+    "eslint": ">=8.0.0",
+    "eslint-plugin-jest": "^28.0.0",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-vue": "^10.0.0"
+  },
+  "peerDependenciesMeta": {
+    "@babel/eslint-parser": {
+      "optional": true
+    },
+    "@babel/eslint-plugin": {
+      "optional": true
+    },
+    "@eslint-react/eslint-plugin": {
+      "optional": true
+    },
+    "@vitest/eslint-plugin": {
+      "optional": true
+    },
+    "angular-eslint": {
+      "optional": true
+    },
+    "angular-eslint-template-parser": {
+      "optional": true
+    },
+    "eslint-plugin-jest": {
+      "optional": true
+    },
+    "eslint-plugin-react-hooks": {
+      "optional": true
+    },
+    "eslint-plugin-vue": {
+      "optional": true
+    }
   },
   "dependencies": {
     "@1stg/config": "^1.0.4",
-    "@babel/eslint-parser": "^7.27.0",
-    "@babel/eslint-plugin": "^7.27.0",
     "@eslint-community/eslint-plugin-eslint-comments": "^4.5.0",
-    "@eslint-react/eslint-plugin": "^1.46.0",
     "@eslint/js": "^9.24.0",
     "@pkgr/utils": "^3.1.1",
     "@typescript-eslint/eslint-plugin": "^8.29.1",
     "@typescript-eslint/parser": "^8.29.1",
-    "@vitest/eslint-plugin": "^1.1.42",
-    "angular-eslint": "^19.3.0",
-    "angular-eslint-template-parser": "^0.1.3",
     "eslint-config-prettier": "^10.1.2",
     "eslint-import-resolver-typescript": "^4.3.2",
     "eslint-plugin-css": "^0.11.0",
     "eslint-plugin-import-x": "^4.10.3",
-    "eslint-plugin-jest": "^28.11.0",
     "eslint-plugin-jsdoc": "^50.6.9",
     "eslint-plugin-jsonc": "^2.20.0",
     "eslint-plugin-markup": "^1.0.0",
@@ -48,13 +79,11 @@
     "eslint-plugin-n": "^17.17.0",
     "eslint-plugin-prettier": "^5.2.6",
     "eslint-plugin-promise": "^7.2.1",
-    "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-regexp": "^2.7.0",
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "eslint-plugin-sonarjs": "^3.0.2",
     "eslint-plugin-toml": "^0.12.0",
     "eslint-plugin-unicorn": "^58.0.0",
-    "eslint-plugin-vue": "^10.0.0",
     "eslint-plugin-yml": "^1.17.0",
     "globals": "^16.0.0",
     "typescript-eslint": "^8.29.1"
